error handling. simple test

+++ index.js
@@ -11,6 +11,28 @@ const uniqueId = require('./utils/unique-id')
 const waitForFile = require('./utils/wait-for-file')
 
 /**
+ * @typedef {import('net').Socket} Socket
+ * @typedef {import('child_process').ChildProcess} ChildProcess
+ * @typedef {import('p-defer').DeferredPromise} DeferredPromise
+ * @typedef {import('./utils/evented').EventCallback} EventCallback
+ */
+
+/**
+ * @typedef Request
+ * @property {string} packet
+ * @property {DeferredPromise} deferred
+ * @property {boolean} sent
+ */
+
+/**
+ * @typedef Response
+ * @property {number} request_id
+ * @property {string} event
+ * @property {string} error
+ * @property {*} [data]
+ */
+
+/**
  * @returns {string}
  */
 function generatePipeName () {
@@ -44,49 +66,65 @@ class Mpv {
   /**
    * @param {object} [options]
    * @param {string} [options.exec]
-   * @param {string} [options.pipeName]
+   * @param {string} [options.execPipe]
+   * @param {number} [options.execTimeout]
    */
   constructor (options = {}) {
-    const {
-      exec = findExecutable(),
-      pipeName = generatePipeName()
-    } = options
+    this.exec = options.exec || findExecutable()
+    this.execPipe = options.execPipe || generatePipeName()
+    this.execTimeout = options.execTimeout || 60000
 
     this.events = new Evented()
     this.observables = new Evented()
+
+    /** @type {Map<number, Request>} */
     this.requests = new Map()
+
+    /** @type {string} */
     this.buffer = ''
+
+    this.pipeWatcher = null
+
+    /** @type {?ChildProcess} */
+    this.cp = null
+
+    /** @type {?Socket} */
+    this.sock = null
+
+    /** @type {boolean} */
+    this.closed = false
 
     this.on('property-change', ({ name, data }) => {
       this.observables.emit(name, data)
     })
 
-    this._init({ exec, pipeName })
+    this._init()
   }
 
   /**
    * @private
-   * @param {object} options
-   * @param {string} options.exec
-   * @param {string} options.pipeName
    */
-  _init ({ exec, pipeName }) {
-    fs.unlink(pipeName, (err) => {
-      if (err && err.code !== 'ENOENT') {
-        console.error(new Error('could not remove pipe file'))
+  _init () {
+    fs.unlink(this.execPipe, (error) => {
+      if (this.closed) {
         return
       }
 
-      this._spawn({ exec, pipeName })
+      if (error && error.code !== 'ENOENT') {
+        this._fatalError(new Error('Could not remove pipe file'))
+        return
+      }
 
-      waitForFile({
-        filename: pipeName,
-        timeout: 5000,
+      this._spawn()
+
+      this.pipeWatcher = waitForFile({
+        filename: this.execPipe,
+        timeout: this.execTimeout,
         onsuccess: () => {
-          this._connect({ pipeName })
+          this._connect()
         },
         ontimeout: () => {
-          console.error(new Error('pipe watcher timed out'))
+          this._fatalError(new Error('Pipe timed out'))
         }
       })
     })
@@ -94,36 +132,47 @@ class Mpv {
 
   /**
    * @private
-   * @param {object} options
-   * @param {string} options.exec
-   * @param {string} options.pipeName
    */
-  _spawn ({ exec, pipeName }) {
-    this.cp = spawn(exec, [
+  _spawn () {
+    if (this.closed) {
+      return
+    }
+
+    this.cp = spawn(this.exec, [
       '--idle',
       '--really-quiet',
       '--no-terminal',
       '--no-video',
       '--no-config',
-      '--input-ipc-server', pipeName
+      '--input-ipc-server', this.execPipe
     ], {
       detached: false,
       windowsHide: true
+    })
+
+    this.cp.on('close', () => {
+      this._fatalError(new Error('MPV executable was unexpectably terminated'))
     })
   }
 
   /**
    * @private
-   * @param {object} options
-   * @param {string} options.pipeName
    */
-  async _connect ({ pipeName }) {
-    this.sock = net.connect(pipeName, () => {
+  async _connect () {
+    if (this.closed) {
+      return
+    }
+
+    this.sock = net.connect(this.execPipe, () => {
       this._flush()
     })
 
-    this.sock.on('error', (...args) => {
-      console.error(...args)
+    this.sock.on('close', () => {
+      this._fatalError(new Error('Socket connection was unexpectably terminated'))
+    })
+
+    this.sock.on('error', (error) => {
+      this._fatalError(error)
     })
 
     this.sock.on('data', data => {
@@ -141,26 +190,54 @@ class Mpv {
   }
 
   /**
-   * 
+   * @private
+   * @param {Error} error
+   */
+  _fatalError (error) {
+    this.close()
+    this.events.emit('error', error)
+  }
+
+  /**
+   *
    */
   close () {
-    this.sock.end()
-    this.cp.kill()
+    this.closed = true
+
+    if (this.pipeWatcher) {
+      this.pipeWatcher.stop()
+      this.pipeWatcher = null
+    }
+
+    if (this.sock) {
+      this.sock.end()
+      this.sock.destroy()
+      this.sock = null
+    }
+
+    if (this.cp) {
+      this.cp.kill()
+      this.cp = null
+    }
+
+    fs.unlink(this.execPipe, () => {})
   }
 
   /**
    * @param {string} command
-   * @returns {Promise<string>}
+   * @param {...*} args
+   * @returns {Promise<*>}
    */
   async command (command, ...args) {
     const id = uniqueId()
     const deferred = defer()
 
-    const packet = {
+    const packet = JSON.stringify({
       request_id: id,
       command: [command, ...args]
-    }
+    })
 
+    /** @type {Request} */
     const request = { deferred, packet, sent: false }
     this.requests.set(id, request)
 
@@ -178,15 +255,15 @@ class Mpv {
 
     for (const [, request] of this.requests) {
       if (request.sent) continue
-      const data = JSON.stringify(request.packet)
-      log('receives', data)
-      this.sock.write(data + '\n')
+      log('receives', request.packet)
+      this.sock.write(request.packet + '\n')
       request.sent = true
     }
   }
 
   /**
    * @private
+   * @param {Response} response
    */
   _processResponse (response) {
     const { request_id, event, error, ...rest } = response
@@ -210,13 +287,15 @@ class Mpv {
 
   /**
    * @param {string} eventName
+   * @param {EventCallback} callback
    */
-  on (eventName, ...args) {
-    return this.events.on(eventName, ...args)
+  on (eventName, callback) {
+    return this.events.on(eventName, callback)
   }
 
   /**
    * @param {string} propertyName
+   * @param {EventCallback} callback
    */
   async observe (propertyName, callback) {
     const id = uniqueId()
@@ -232,7 +311,7 @@ class Mpv {
 
   /**
    * @param {string} propertyName
-   * @returns {Promise<string>}
+   * @returns {Promise<*>}
    */
   async get (propertyName) {
     return this.command('get_property', propertyName)
